requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
langgraph==0.2.28
langchain-core==0.3.15
pydantic==2.9.2
python-dotenv==1.0.1
